requirements.txt
streamlit
langchain
langchain_community
python-dotenv
langchain_core
langserve
fastapi
uvicorn
sse_starlette